Updates to the test setup and package names
