+++ examples/example_cel_native_structs_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package examples
+package examples_test
 
 import (
 	"fmt"
@@ -20,23 +20,13 @@ import (
 	"reflect"
 
 	"cel.dev/cel-go/cel"
+	"cel.dev/cel-go/examples"
 	"cel.dev/cel-go/ext"
 )
 
-type User struct {
-	Name  string   `json:"name"`
-	Age   int      `json:"age"`
-	Roles []string `json:"roles"`
-}
-
-type Account struct {
-	ID        int64  `cel:"id"`
-	OwnerName string `cel:"owner"`
-}
-
 // Example_cel_NativeTypes showcases evaluating Go native structs using ext.NativeTypes() with json struct tags
 func Example_cel_NativeTypes() {
-	u := User{
+	u := examples.User{
 		Name:  "Alice",
 		Age:   30,
 		Roles: []string{"admin", "editor"},
@@ -52,7 +42,7 @@ func Example_cel_NativeTypes() {
 	for _, expr := range exprs {
 		prg, err := cel.Compile(expr,
 			cel.Variable("user", cel.ObjectType("examples.User")),
-			ext.NativeTypes(reflect.TypeOf(User{}), ext.ParseStructTag("json")),
+			ext.NativeTypes(reflect.TypeFor[examples.User](), ext.ParseStructTag("json")),
 		)
 		if err != nil {
 			log.Fatalf("cel.Compile() error for %q: %v", expr, err)
@@ -73,14 +63,14 @@ func Example_cel_NativeTypes() {
 
 // Example_cel_NativeTypes_structTags showcases mapping struct field names via cel tags
 func Example_cel_NativeTypes_structTags() {
-	acc := Account{
+	acc := examples.Account{
 		ID:        1001,
 		OwnerName: "Bob",
 	}
 
 	prg, err := cel.Compile(`acc.owner == "Bob" && acc.id == 1001`,
 		cel.Variable("acc", cel.ObjectType("examples.Account")),
-		ext.NativeTypes(reflect.TypeFor[Account](), ext.ParseStructTags(true)),
+		ext.NativeTypes(reflect.TypeFor[examples.Account](), ext.ParseStructTags(true)),
 	)
 	if err != nil {
 		log.Fatalf("cel.Compile() error: %v", err)
