+++ examples/example_cel_type_conversions_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package examples
+package examples_test
 
 import (
 	"fmt"
